#step2 [refactor] 반환 자료형 변경
- 반환 자료형은 int로 하고, 실제로 rate를 구할때에만 double로 변환해서 계산하도록 수정

+++ src/main/java/lotto/model/Lottos.java
@@ -17,7 +17,7 @@ public class Lottos {
         return this.lottos.size();
     }
 
-    public double getBuyingMoney() {
+    public int getBuyingMoney() {
         return Lotto.PRICE * this.lottos.size();
     }
